Fixed unit tests

+++ Logger/FileLogger.cs
@@ -7,25 +7,31 @@ namespace Logger
 {
     public class FileLogger : BaseLogger
     {
-        public object ClassName { get; set; }
-        public DateTime dateTime;
-        public FileStream file;
-        public StreamWriter writer;
+        public object? ClassName { get; set; }
+        public DateTime? dateTime;
+        public FileStream? file;
+        public StreamWriter? writer;
 
 
-        public FileLogger(string Path, object Name)
+        public FileLogger(string? Path, object? Name)
         {
             ClassName = Name;
-            file = File.Open(Path, FileMode.Open, FileAccess.Write, FileShare.None);
-            writer = new StreamWriter(file);
+            if (File.Exists(Path))
+            {
+                file = File.Open(Path, FileMode.Append, FileAccess.Write, FileShare.None);
+                writer = new StreamWriter(file);
+            }
         }
         public override void Log(LogLevel logLevel, string message)
         {
             dateTime = DateTime.Now;
             string appendLine = dateTime + " " + ClassName + " "
                 + logLevel + " " + message;
-            writer.WriteLine(appendLine);
-            Console.WriteLine(appendLine);
+            if (writer != null)
+            {
+                writer.WriteLine(appendLine);
+                Console.WriteLine(appendLine);
+            }
         }
     }
 }
